requirements.txt
discord.py
requests
python-dotenv
google-generative-ai
